Arreglar bug al eliminar propiedad inexistente

+++ src/controllers/propertyController.js
@@ -67,7 +67,6 @@ export const updateProperty = async (req, res) => {
     }
 }
 
-// TODO Fix Bug When Trying To Delete A Property That Doesn't Exist
 export const deleteProperty = async (req, res) => {
     try {
         const { errors } = validationResult(req);
@@ -79,6 +78,14 @@ export const deleteProperty = async (req, res) => {
         }
 
         const { id } = req.params;
+        // Check If Property Exist
+        const property = await Property.findByPk(id);
+        if (property === null) {
+            return res.status(404).json({
+                msg: "property not registered in the database"
+            });
+        }
+
         await Property.destroy({
             where: {
                 id
